Move Places to population package
