+++ src/main/java/uk/co/ramp/covid/simulation/population/Child.java
@@ -2,9 +2,6 @@ package uk.co.ramp.covid.simulation.population;
 
 import uk.co.ramp.covid.simulation.CovidParameters;
 import uk.co.ramp.covid.simulation.DailyStats;
-import uk.co.ramp.covid.simulation.place.CommunalPlace;
-import uk.co.ramp.covid.simulation.place.Places;
-import uk.co.ramp.covid.simulation.place.School;
 
 public class Child extends Person {
