+++ src/main/java/uk/co/ramp/covid/simulation/population/Infant.java
@@ -2,9 +2,6 @@ package uk.co.ramp.covid.simulation.population;
 
 import uk.co.ramp.covid.simulation.CovidParameters;
 import uk.co.ramp.covid.simulation.DailyStats;
-import uk.co.ramp.covid.simulation.place.CommunalPlace;
-import uk.co.ramp.covid.simulation.place.Nursery;
-import uk.co.ramp.covid.simulation.place.Places;
 
 public class Infant extends Person {
