+++ src/main/java/uk/co/ramp/covid/simulation/population/Pensioner.java
@@ -2,7 +2,6 @@ package uk.co.ramp.covid.simulation.population;
 
 import uk.co.ramp.covid.simulation.CovidParameters;
 import uk.co.ramp.covid.simulation.DailyStats;
-import uk.co.ramp.covid.simulation.place.Places;
 
 public class Pensioner extends Person {
